Split schemas into separate files
