+++ wagtail/api/v3/schemas/__init__.py
@@ -1,0 +1,11 @@
+from .base import ContentTypeSummarySchema
+from .pages import BasePageSchema, PageMetaSchema
+from .sites import SiteInputSchema, SiteSchema
+
+__all__ = [
+    "ContentTypeSummarySchema",
+    "BasePageSchema",
+    "PageMetaSchema",
+    "SiteInputSchema",
+    "SiteSchema",
+]
