+++ wagtail/api/v3/schemas/base.py
@@ -1,0 +1,6 @@
+from ninja import Schema
+
+
+class ContentTypeSummarySchema(Schema):
+    name: str
+    label: str
