+++ wagtail/api/v3/schemas/pages.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 from django.urls import reverse
 from django.urls.exceptions import NoReverseMatch
-from ninja import Field, Schema
+from ninja import Schema
 
 from wagtail.api.v2.utils import get_full_url
 from wagtail.models import Page
@@ -43,27 +43,3 @@ class BasePageSchema(Schema):
             slug=obj.slug,
             first_published_at=obj.first_published_at,
         )
-
-
-class SiteSchema(Schema):
-    id: int
-    hostname: str
-    port: int
-    site_name: str
-    root_page_id: int
-    is_default_site: bool
-
-
-class SiteInputSchema(Schema):
-    hostname: str
-    port: int = 80
-    site_name: str = ""
-    # Accept root_page_id in the input schema for consistency with the output,
-    # but use root_page so it can be accepted by ModelForm.
-    root_page: int = Field(..., alias="root_page_id")
-    is_default_site: bool = False
-
-
-class ContentTypeSummarySchema(Schema):
-    name: str
-    label: str
